Fix Hell

+++ Content.Server/_Starlight/Shadekin/ShadekinSystem.Brighteye.cs
@@ -197,7 +197,7 @@ public sealed partial class ShadekinSystem : EntitySystem
 
         if (component.Energy >= cost)
         {
-            component.Energy -= cost;
+            component.Energy -= (int)cost;
             Dirty(uid, component);
         }
         else
